% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{build_palette}
\alias{build_palette}
\title{Generate a palette of related colors around a central color}
\usage{
build_palette(
  central_color,
  n_colors = 10,
  hue_range = 0.1,
  sat_range = 0.4,
  val_range = 0.45
)
}
\arguments{
\item{central_color}{a single color value around which to build the palette}

\item{n_colors}{the number of colors to return in the primary palette (max = 100, default = 10)}

\item{hue_range}{how much variation in hue is allowed in the palette (range = 0 to 1, default = 0.1)}

\item{val_range}{how much variation in value is allowed in the palette (range = 0 to 1, default = 0.45)}
}
\value{
a list of palette results, containing:
\itemize{
\item palette: a character vector with a 10-color palette realted to the central_color.
\item palette_plot: a ggplot2 object showing the 10 colors in palette.
\item colorset: A character vector with the full set of 100 colors from which the palette was selected.
\item colorset_plot: A ggplot2 object displaying all 100 colors in the colorset.
\item weave: a ggplot2 object generated with weave_plot() showing the intersection of all of the collors in the palette.
}
}
\description{
Generate a palette of related colors around a central color
}
