% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing.R
\name{color_mean}
\alias{color_mean}
\title{Compute the mean of multiple colors in RGB space}
\usage{
color_mean(color_vec)
}
\arguments{
\item{color_vec}{A vector of hex or R colors}
}
\value{
The mean of the colors as a character hex color (e.g. "#FFFFFF")
}
\description{
Compute the mean of multiple colors in RGB space
}
