% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustment.R
\name{desaturate_color}
\alias{desaturate_color}
\title{Desaturate colors by reducing Saturation in HSV space}
\usage{
desaturate_color(color_vec, desat_frac = 0.5, sat_floor = 0.1)
}
\arguments{
\item{color_vec}{a vector of hex or R colors}

\item{deval_frac}{numeric (0 to 1) - the amount to subtract from Saturation in HSV}

\item{val_floor}{numeric (0 to 1) - a minimum Saturation to retain for each color}
}
\value{
a character vector of desaturated hex colors
}
\description{
Desaturate colors by reducing Saturation in HSV space
}
