% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustment.R
\name{fade_color}
\alias{fade_color}
\title{Fade colors by reducing Value in HSV space}
\usage{
fade_color(color_vec, deval_frac = 0.5, val_floor = 0.1)
}
\arguments{
\item{color_vec}{a vector of hex or R colors}

\item{deval_frac}{numeric (0 to 1) - the amount to subtract from Value in HSV}

\item{val_floor}{numeric (0 to 1) - a minimum Value to retain for each color}
}
\value{
a character vector of faded hex colors
}
\description{
Fade colors by reducing Value in HSV space
}
