% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{force_all}
\alias{force_all}
\title{Force a list}
\usage{
force_all(...)
}
\arguments{
\item{...}{Parameters to change to a list.}
}
\value{
a list object
}
\description{
This is a helper function for varibow_pal() and is not exported.
}
