% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing.R
\name{merge_nearest_colors}
\alias{merge_nearest_colors}
\title{Merge colors based on similarity and frequency}
\usage{
merge_nearest_colors(color_vec, k = 12)
}
\arguments{
\item{color_vec}{A vector of hex or R colors}

\item{k}{The number of colors to retain}
}
\value{
a vector of merged colors with equal length to color_vec.
}
\description{
Merge colors based on similarity and frequency
}
