% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_varibow}
\alias{scale_fill_varibow}
\title{Varibow colorscale for ggplot2}
\usage{
scale_fill_varibow(..., alpha = 1, aesthetics = "fill")
}
\arguments{
\item{...}{Parameters passed to ggplot2::discrete_scale()}

\item{alpha}{Alpha value}

\item{aesthetics}{Passed to ggplot2::discrete_scale(). Default is "fill".}
}
\value{
A ggplot2 discrete_scale() result.
}
\description{
Varibow colorscale for ggplot2
}
