% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{varibow_pal}
\alias{varibow_pal}
\title{Generate a varibow palette for ggplot}
\usage{
varibow_pal(alpha = 1)
}
\arguments{
\item{alpha}{The alpha for the palette. if NULL, alpha hex values are trimmed. Default = 1.}
}
\value{
a varibow palette
}
\description{
This is a helper function for scale_color_varibow() and is not exported.
}
